requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.88.0
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
jmespath==1.0.1
psycopg2==2.9.6
pydantic==1.10.3
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
